requirements.txt
streamlit==1.12.0
torch
transformers
diffusers
Pillow
